+++ CHANGELOG.md
@@ -7,8 +7,10 @@
 ### 新特性
 * 【core】新增PinyinComparator、CollUtil新增sortByPinyin（感谢@【帝都】宁静）
 * 【json】JSONUtil增加xmlToJson方法
+* 【poi】 ExcelWriter增加setColumnWidth和setRowHeight方法
 
 ### Bug修复
+* 【json】修复JSONArray中addAll加入两次的bug（感谢@【天津】〓下页）
 
 -------------------------------------------------------------------------------------------------------------